backend/requirements.txt
asgiref==3.8.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
git+https://github.com/Juunsik/dj-rest-auth.git
Django==4.2
django-allauth==0.62.1
django-dotenv==1.4.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
idna==3.7
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
PyJWT==2.8.0
PyYAML==6.0.1
referencing==0.35.1
requests==2.31.0
rpds-py==0.18.1
sqlparse==0.5.0
uritemplate==4.1.1
urllib3==2.2.1
